+++ src/envguard/hook/__main__.py
@@ -1,6 +1,0 @@
-"""Make envguard.hook runnable as a module: python3 -m envguard.hook"""
-
-from . import main
-
-if __name__ == "__main__":
-    main()